feat(reporting): implement monthly summary read repository with EF

+++ src/FamilyFinances.Application/Reporting/Dtos/AccountTotalsDto.cs
@@ -13,6 +13,6 @@ public sealed record AccountTotalItemDto(
     string AccountName,
     AccountNature AccountNature,
     AccountKind AccountKind,
-    decimal NetChange,
+    long NetChange,
     int TransactionsCount
 );
