+++ src/FamilyFinances.Application/Reporting/Dtos/MonthlySummaryDto.cs
@@ -3,8 +3,8 @@ namespace FamilyFinances.Application.Reporting.Dtos;
 public sealed record MonthlySummaryDto(
     int Year,
     int Month,
-    decimal IncomeTotal,
-    decimal ExpenseTotal,
-    decimal Net,
+    long IncomeTotal,
+    long ExpenseTotal,
+    long Net,
     int TransactionsCount
 );
